refactor file folder structure
