+++ nanome_docking/menus/icons/DarkOval.png
