+++ nanome_docking/menus/icons/ligand_gray.png
