+++ nanome_docking/menus/icons/receptor_gray.png
