+++ nanome_docking/menus/icons/receptor_white.png
